requirements.txt
Django>=5.2
djangorestframework>=3.15.0
django-cors-headers==4.3.1
django-filter>=24.0
Pillow>=10.3.0
python-decouple==3.8
gunicorn
